refactor: rename unit tests
