+++ src/clients/actor-client.test.ts
@@ -2,8 +2,8 @@ import type { RunClient } from 'apify-client';
 import { ActorClient, ApifyClient } from 'apify-client';
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 
-import { getClientContext } from '../__test-helpers__/context.js';
-import { createActorRunMock } from '../__test-helpers__/mocks.js';
+import { getClientContext } from '../__unit__/context.js';
+import { createActorRunMock } from '../__unit__/mocks.js';
 import type { ClientContext } from '../context/client-context.js';
 import type { RunSource } from '../entities/run-source.js';
 import type { ExtActorClient } from './actor-client.js';
