+++ src/clients/apify-client.test.ts
@@ -1,8 +1,8 @@
 import { RunClient } from 'apify-client';
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 
-import { getClientContext } from '../__test-helpers__/context.js';
-import { createActorRunMock } from '../__test-helpers__/mocks.js';
+import { getClientContext } from '../__unit__/context.js';
+import { createActorRunMock } from '../__unit__/mocks.js';
 import type { ClientContext } from '../context/client-context.js';
 import { RunSource } from '../entities/run-source.js';
 import { ExtActorClient } from './actor-client.js';
